no heuristic code

+++ offline/core/service.py
@@ -9,7 +9,7 @@ from sqlalchemy import Column, Integer, ForeignKey
 from sqlalchemy import and_
 from sqlalchemy.orm import relationship
 
-from ..core.service_topo import ServiceTopo
+from ..core.service_topo import ServiceTopoHeuristic, ServiceTopoFull
 from ..core.sla import Sla, SlaNodeSpec
 from ..core.solver import solve
 from ..time.persistence import ServiceNode, ServiceEdge, Base, service_to_sla
@@ -21,9 +21,9 @@ RESULTS_FOLDER = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../r
 
 def f(x):
     session = Session()
-    slasIDS, vhg_count, vcdn_count = x
+    slasIDS, vhg_count, vcdn_count, use_heuristic = x
     service = Service(slasIDS=slasIDS, serviceSpecFactory=ServiceSpecFactory, vhg_count=vhg_count,
-                      vcdn_count=vcdn_count)
+                      vcdn_count=vcdn_count, use_heuristic=use_heuristic)
     session.add(service)
     return service.id
 
@@ -89,8 +89,8 @@ class Service(Base):
         merged_new = self.__get_merged_sla(self.slas)
 
         # create the service topology from the new merged sla
-        self.topo = ServiceTopo(sla=merged_new, vhg_count=self.vhg_count,
-                                vcdn_count=self.vcdn_count, hint_node_mappings=self.mapping.node_mappings)
+        self.topo = ServiceTopoHeuristic(sla=merged_new, vhg_count=self.vhg_count,
+                                         vcdn_count=self.vcdn_count, hint_node_mappings=self.mapping.node_mappings)
 
         # retreive info on the new service.
         edges_from_new_topo = {(node_1, node_2): bw for node_1, node_2, bw in self.topo.dump_edges()}
@@ -186,7 +186,7 @@ class Service(Base):
 
     @classmethod
     def get_optimal(cls, slas, serviceSpecFactory=ServiceSpecFactory, max_vhg_count=10, max_vcdn_count=10,
-                    threads=multiprocessing.cpu_count() - 1, remove_service=True):
+                    threads=multiprocessing.cpu_count() - 1, remove_service=True, use_heuristic=True):
         session = Session()
         threadpool = ThreadPool(threads)
         thread_param = []
@@ -203,10 +203,10 @@ class Service(Base):
 
         for vhg_count in range(1, max_vhg_count + 1, ):
             for vcdn_count in range(1, min(vhg_count, max_vcdn_count) + 1):
-                thread_param.append(([sla.id for sla in slas], vhg_count, vcdn_count))
+                thread_param.append(([sla.id for sla in slas], vhg_count, vcdn_count, use_heuristic))
 
-        services = threadpool.map(f, thread_param)
-        # services = [f(x) for x in thread_param]
+        #services = threadpool.map(f, thread_param)
+        services = [f(x) for x in thread_param]
         services = session.query(Service).filter(Service.id.in_(services)).all()
 
         for service in services:
@@ -234,9 +234,13 @@ class Service(Base):
 
         return best_service
 
-    def __init__(self, slasIDS, serviceSpecFactory=ServiceSpecFactory, vhg_count=1, vcdn_count=1):
+    def __init__(self, slasIDS, serviceSpecFactory=ServiceSpecFactory, vhg_count=1, vcdn_count=1, use_heuristic=True):
         session = Session()
-        # print("7777777777777777 %s" % str(session))
+
+        if use_heuristic:
+            serviceTopoClass = ServiceTopoHeuristic
+        else:
+            serviceTopoClass = ServiceTopoFull
 
         self.slas = session.query(Sla).filter(Sla.id.in_(slasIDS)).all()
         self.vhg_count = vhg_count
@@ -244,7 +248,8 @@ class Service(Base):
         self.merged_sla = self.__get_merged_sla(self.slas)
         self.serviceSpecFactory = serviceSpecFactory
 
-        self.topo = {sla: ServiceTopo(sla=sla, vhg_count=vhg_count, vcdn_count=vcdn_count, hint_node_mappings=None) for
+        self.topo = {sla: serviceTopoClass(sla=sla, vhg_count=vhg_count, vcdn_count=vcdn_count, hint_node_mappings=None)
+                     for
                      sla in
                      [self.merged_sla]}
         # self.slas}
@@ -269,45 +274,48 @@ class Service(Base):
                 self.serviceEdges.append(sedge)
             session.flush()
 
-        # create temp mapping for vhg<->vcdn hints
-        assert self.id is not None
-        self.__solve(path=str(self.id))
-        session.flush()
 
-        if self.mapping is not None:
-            self.topo = {sla: ServiceTopo(sla=sla, vhg_count=vhg_count, vcdn_count=vcdn_count,
-                                          hint_node_mappings=self.mapping.node_mappings) for sla in [self.merged_sla]}
-
-            # add the CDN Edges to the graph
-            for sla in [self.merged_sla]:
-
-                for node_1, node_2, bandwidth in self.topo[sla].getServiceCDNEdges():
-                    snode_1 = session.query(ServiceNode).filter(
-                        and_(ServiceNode.sla_id == sla.id, ServiceNode.service_id == self.id,
-                             ServiceNode.name == node_1)).one()
-
-                    snode_2 = session.query(ServiceNode).filter(
-                        and_(ServiceNode.sla_id == sla.id, ServiceNode.service_id == self.id,
-                             ServiceNode.name == node_2)).one()
-
-                    sedge = ServiceEdge(node_1=snode_1, node_2=snode_2, bandwidth=bandwidth, sla_id=sla.id)
-                    session.add(sedge)
-                    self.serviceEdges.append(sedge)
-                session.flush()
-
-            session.delete(self.mapping)
-            session.flush()
+        if use_heuristic:
+            # create temp mapping for vhg<->vcdn hints
+            assert self.id is not None
             self.__solve(path=str(self.id))
             session.flush()
 
-    def __solve(self, path="."):
+            if self.mapping is not None:
+                self.topo = {sla: ServiceTopoHeuristic(sla=sla, vhg_count=vhg_count, vcdn_count=vcdn_count,
+                                                       hint_node_mappings=self.mapping.node_mappings) for sla in
+                             [self.merged_sla]}
+
+                # add the CDN Edges to the graph
+                for sla in [self.merged_sla]:
+
+                    for node_1, node_2, bandwidth in self.topo[sla].getServiceCDNEdges():
+                        snode_1 = session.query(ServiceNode).filter(
+                            and_(ServiceNode.sla_id == sla.id, ServiceNode.service_id == self.id,
+                                 ServiceNode.name == node_1)).one()
+
+                        snode_2 = session.query(ServiceNode).filter(
+                            and_(ServiceNode.sla_id == sla.id, ServiceNode.service_id == self.id,
+                                 ServiceNode.name == node_2)).one()
+
+                        sedge = ServiceEdge(node_1=snode_1, node_2=snode_2, bandwidth=bandwidth, sla_id=sla.id)
+                        session.add(sedge)
+                        self.serviceEdges.append(sedge)
+                    session.flush()
+
+                session.delete(self.mapping)
+                session.flush()
+        self.__solve(path=str(self.id),use_heuristic=use_heuristic)
+        session.flush()
+
+    def __solve(self, path=".",use_heuristic=True):
         '''
         Solve the service according to specs
         :return: nothing, service.mapping may be initialized with an actual possible mapping
         '''
         session = Session()
         if len(self.slas) > 0:
-            solve(self, self.slas[0].substrate, path)
+            solve(self, self.slas[0].substrate, path,use_heuristic)
             if self.mapping is not None:
                 session.add(self.mapping)
 
@@ -353,7 +361,7 @@ class Service(Base):
                 postfix = "%d_%d" % (self.id, sla.id)
                 topo = self.topo[sla]
                 for snode_id, cpu, bw in topo.getServiceNodes():
-                    f.write("%s_%s %lf %lf\n" % (snode_id, postfix, cpu,bw))
+                    f.write("%s_%s %lf %lf\n" % (snode_id, postfix, cpu, bw))
                     # sys.stdout.write("%s_%s %lf\n" % (snode_id, postfix, cpu))
 
 
@@ -368,8 +376,8 @@ class Service(Base):
         with open(os.path.join(RESULTS_FOLDER, path, "starters.nodes.data"), mode) as f:
             for sla in slas:
                 postfix = "%d_%d" % (self.id, sla.id)
-                for s, topo in self.topo[sla].get_Starters():
-                    f.write("%s_%s %s\n" % (s, postfix, topo))
+                for s, topo, bw in self.topo[sla].get_Starters():
+                    f.write("%s_%s %s %lf\n" % (s, postfix, topo,bw))
 
         # write the names of the VHG Nodes
         with open(os.path.join(RESULTS_FOLDER, path, "VHG.nodes.data"), mode) as f:
